Change the length type of stringops from int to size_t.

+++ src/utils/small_stringops.h
@@ -24,13 +24,13 @@ namespace ufal {
 namespace utils {
 
 // Declarations
-inline bool small_memeq(const void* a, const void* b, int len);
-inline int small_memcmp(const void* a, const void* b, int len);
-inline void small_memcpy(void* dest, const void* src, int len);
-inline int small_strnchrpos(const char* str, char c, int len);
+inline bool small_memeq(const void* a, const void* b, size_t len);
+inline int small_memcmp(const void* a, const void* b, size_t len);
+inline void small_memcpy(void* dest, const void* src, size_t len);
+inline size_t small_strnchrpos(const char* str, char c, size_t len);
 
 // Definitions
-inline bool small_memeq(const void* a_void, const void* b_void, int len) {
+inline bool small_memeq(const void* a_void, const void* b_void, size_t len) {
   const char* a = (const char*)a_void;
   const char* b = (const char*)b_void;
 
@@ -40,7 +40,7 @@ inline bool small_memeq(const void* a_void, const void* b_void, int len) {
   return true;
 }
 
-inline int small_memcmp(const void* a_void, const void* b_void, int len) {
+inline int small_memcmp(const void* a_void, const void* b_void, size_t len) {
   const char* a = (const char*)a_void;
   const char* b = (const char*)b_void;
 
@@ -50,7 +50,7 @@ inline int small_memcmp(const void* a_void, const void* b_void, int len) {
   return 0;
 }
 
-inline void small_memcpy(void* dest_void, const void* src_void, int len) {
+inline void small_memcpy(void* dest_void, const void* src_void, size_t len) {
   char* dest = (char*)dest_void;
   const char* src = (const char*)src_void;
 
@@ -58,8 +58,8 @@ inline void small_memcpy(void* dest_void, const void* src_void, int len) {
     *dest++ = *src++;
 }
 
-inline int small_strnchrpos(const char* str, char c, int len) {
-  int pos = 0;
+inline size_t small_strnchrpos(const char* str, char c, size_t len) {
+  size_t pos = 0;
   for (; len--; str++, pos++)
     if (*str == c)
       return pos;
